Final Project/requirements.txt
# This file should list all of the Python dependencies your project needs.
# You can generate it by running `pip freeze > requirements.txt` in your project's virtual environment.

customtkinter
Pillow
matplotlib
tensorflow
keras
scikit-learn
pandas
numpy
